1028 리스폰지역으로 돌아오는 기능 추가, 보스타워 파괴시 게임오버

+++ Source/LOBG/Battle/BattlePS.h
@@ -32,7 +32,7 @@ public:
 		float PlayerMoneyData;
 
 	UPROPERTY(BlueprintReadWrite, EditAnywhere)
-		float PlayerLevel;
+		int PlayerLevel;
 };
 /**
  *
